move note command exceptions to own package
