+++ src/main/java/seedu/lifeasier/commands/Command.java
@@ -1,6 +1,5 @@
 package seedu.lifeasier.commands;
 
-import seedu.lifeasier.exceptions.TitleNotFoundException;
 import seedu.lifeasier.storage.FileStorage;
 import seedu.lifeasier.tasks.TaskList;
 import seedu.lifeasier.ui.Ui;
@@ -20,8 +19,7 @@ public abstract class Command {
     /**
      * Executes the user command.
      */
-    public abstract void execute(Ui ui, NoteList notes, TaskList tasks, FileStorage storage) throws
-            TitleNotFoundException;
+    public abstract void execute(Ui ui, NoteList notes, TaskList tasks, FileStorage storage);
 
     /**
      * Returns status of whether the program has been terminated.
